# Conflicts:
#	src/main/java/com/review/shop/controller/ImageUploadController.java

+++ src/main/java/com/review/shop/dto/orders/OrderItemDTO.java
@@ -2,11 +2,10 @@ package com.review.shop.dto.orders;
 
 import lombok.Data;
 
-
 @Data
 public class OrderItemDTO {
-    private long order_item_id;
-    private long order_id;      // ORDERS insert 후 받아온 ID
+    private int order_item_id; // long -> int 변경
+    private int order_id;      // long -> int 변경
     private int product_id;
     private String prd_name;
     private int product_price;
